[refactor] globalcar에 initalCar추가
